Add navbar, closes #391
